refactor: change the props location
